requirements.txt
fastmcp
pytesseract
pdf2image
tqdm
python-dotenv